Prefer to remove FileUtils for removing folder
Currently, when this is run the folder always exists before
`Spec::Rubygems.setup` has been run in the `before(:suite)` hook.
However, I plan to refactor that to only setup the test environment but
not install any gems.

When I do that, that folder might not exist at this point, so we need
`FileUtils.rm_rf` which correctly handles that.

Also, this method uses `FileUitls` in other places so this change makes
file manipulation in the method more consistent.